tailscale: rename tailscale_test package to tailscale
This allows the reuse of private functions like deviceToMap between the core package and the tests.

Updates tailscale/corp#21867
