+++ tailscale/data_source_4via6_test.go
@@ -1,4 +1,4 @@
-package tailscale_test
+package tailscale
 
 import (
 	"fmt"
