+++ tailscale/data_source_acl_test.go
@@ -1,4 +1,4 @@
-package tailscale_test
+package tailscale
 
 import (
 	"context"
@@ -9,7 +9,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
 	"github.com/tailscale/hujson"
-	"github.com/tailscale/terraform-provider-tailscale/tailscale"
 )
 
 func TestAccTailscaleACL(t *testing.T) {
@@ -22,7 +21,7 @@ func TestAccTailscaleACL(t *testing.T) {
 			{
 				Config: `data "tailscale_acl" "acl" {}`,
 				Check: func(s *terraform.State) error {
-					client := testAccProvider.Meta().(*tailscale.Clients).V2
+					client := testAccProvider.Meta().(*Clients).V2
 					acl, err := client.PolicyFile().Raw(context.Background())
 					if err != nil {
 						return fmt.Errorf("unable to get ACL: %s", err)
