+++ tailscale/resource_acl_test.go
@@ -1,4 +1,4 @@
-package tailscale_test
+package tailscale
 
 import (
 	"context"
