+++ tailscale/resource_dns_split_nameservers_test.go
@@ -1,4 +1,4 @@
-package tailscale_test
+package tailscale
 
 import (
 	"context"
