+++ tailscale/resource_tailnet_settings_test.go
@@ -1,4 +1,4 @@
-package tailscale_test
+package tailscale
 
 import (
 	"context"
