+++ tailscale/resource_webhook_test.go
@@ -1,4 +1,4 @@
-package tailscale_test
+package tailscale
 
 import (
 	"context"
